Add files via upload
I needed the ability to export all detections within a specific hour range from a video camera system.  This will connect to the Unifi Protect Postgresql database, looking for detections on a specific camera.  If the detection falls within a given time frame, it will then convert the detection time from UTC to local time.  From there it will create a folder structure, year/month/day as needed.  It will then export the UBV file to MP4, and export the detection to an MP4 naming the file based on the given camera name, local hour, minute and second.  

It is not well refined code, but this was built for a specific purpose.  Hopefully it will help someone else.